refactor: Extract swipe action performing logic to its own file to reduce cognitive complexity
Sonar notified us that this method had too much cognitive complexity, but to not make the other PR noisy, I fixed the cognitive complexity here instead.

It also removes 100+ lines from the already very big ThreadListFragment, so it's an even better refactor.
